requirements.txt
pandas>=1.3.0
openpyxl>=3.0.7
python-docx>=0.8.11
pyyaml>=6.0
click>=8.0.0
python-docx>=0.8.11
xlrd>=2.0.1
